+++ classes/prefix_WPgutenberg/blocks/core-extensions/image/inspector.js
@@ -38,6 +38,7 @@ export default class Inspector extends Component {
       attributes: {
         addSwiper,
         addPopUp,
+        addPopUpInfo,
         addPopUpPreview,
         addDownloadButton
       },
@@ -53,6 +54,14 @@ export default class Inspector extends Component {
                     label={ __( 'Activate Lightbox', 'devTheme' ) }
                     checked={ addPopUp }
                     onChange={addPopUp => setAttributes({ addPopUp })}
+                />
+              </PanelRow>
+              <PanelRow>
+                <ToggleControl
+                    id="posts-popup-info"
+                    label={ __( 'Load post info inside Lightbox', 'devTheme' ) }
+                    checked={ addPopUpInfo }
+                    onChange={addPopUpInfo => setAttributes({ addPopUpInfo })}
                 />
               </PanelRow>
               <PanelRow>