+++ classes/prefix_WPgutenberg/blocks/posts/index.js
@@ -446,7 +446,7 @@ REGISTER BLOCK
     } )( props => {
       // set values
       const {
-        attributes: { postType, postTaxonomyFilter, postIdFilter, postTaxonomyFilterRelation, postSum, postRepeater, postSortDirection, postSortBy, postTextOne, postTextTwo, postColumns, anchor, postThumb, postSwiper, postPopUp, postPopUpNav, postSortNav, postSortNavOptions, postColumnsSpace, postTaxonomyFilterOptions, postsInsideLoad, postsInsideLoadFirst, postsBreakpoints },
+        attributes: { postType, postTaxonomyFilter, postIdFilter, postTaxonomyFilterRelation, postSum, postRepeater, postSortDirection, postSortBy, postTextOne, postTextTwo, postColumns, anchor, postThumb, postSwiper, postPopUp, postPopUpInfo, postPopUpNav, postSortNav, postSortNavOptions, postColumnsSpace, postTaxonomyFilterOptions, postsInsideLoad, postsInsideLoadFirst, postsBreakpoints },
         attributes,
         className,
         setAttributes,
@@ -522,7 +522,7 @@ REGISTER BLOCK
     } ),
     save: props => {
       const {
-        attributes: { postType, postTaxonomyFilter, postIdFilter, postTaxonomyFilterRelation, postSum, postRepeater, postSortDirection, postSortBy, postTextOne, postTextTwo, postColumns, anchor, postThumb, postSwiper, postPopUp, postPopUpNav, postSortNav, postSortNavOptions, postColumnsSpace, postTaxonomyFilterOptions, postsInsideLoad, postsInsideLoadFirst, postsBreakpoints },
+        attributes: { postType, postTaxonomyFilter, postIdFilter, postTaxonomyFilterRelation, postSum, postRepeater, postSortDirection, postSortBy, postTextOne, postTextTwo, postColumns, anchor, postThumb, postSwiper, postPopUp, postPopUpInfo, postPopUpNav, postSortNav, postSortNavOptions, postColumnsSpace, postTaxonomyFilterOptions, postsInsideLoad, postsInsideLoadFirst, postsBreakpoints },
         attributes
       } = props;